+++ cli/srecli/group/tenant.py
@@ -42,9 +42,9 @@ def describe(ctx: ContextObj, tenant_name, account_number, active, cloud,
     """
     if tenant_name and account_number:
         raise click.ClickException(
-            'Either --tenant_name or --account_number can be given'
+            'Either --tenant_name or --account_number can be provided.'
         )
-    if tenant_name:
+    if tenant_name or account_number:
         return ctx['api_client'].tenant_get(tenant_name or account_number,
                                             customer_id=customer_id)
     return ctx['api_client'].tenant_query(